fix(smrt-svelte): extract types to separate files for TypeScript compatibility
TypeScript cannot re-export types from .svelte files. Move types to dedicated
types.ts files in each component directory:
- commerce/types.ts: InvoiceStatus, InvoiceData, LineItem, UnbilledItem
- display/types.ts: StatusType
- nav/types.ts: FilterOption, Tab
