+++ packages/smrt-svelte/src/components/display/index.ts
@@ -2,9 +2,10 @@
  * Display components - Read-only presentation components
  */
 
+// Export components
 export { default as ConfidenceBadge } from './ConfidenceBadge.svelte';
 export { default as CurrencyDisplay } from './CurrencyDisplay.svelte';
 export { default as DateDisplay } from './DateDisplay.svelte';
-// Re-export types
-export type { StatusType } from './StatusBadge.svelte';
 export { default as StatusBadge } from './StatusBadge.svelte';
+// Re-export types from types file
+export type { StatusType } from './types.js';
