+++ packages/smrt-svelte/src/components/display/types.ts
@@ -1,0 +1,12 @@
+/**
+ * Display component types
+ */
+
+export type StatusType =
+  | 'default'
+  | 'invoice'
+  | 'project'
+  | 'expense'
+  | 'time'
+  | 'compliance'
+  | 'estimate';
