+++ packages/smrt-svelte/src/components/nav/index.ts
@@ -2,8 +2,8 @@
  * Navigation components
  */
 
-// Re-export types
-export type { FilterOption } from './FilterChips.svelte';
+// Export components
 export { default as FilterChips } from './FilterChips.svelte';
-export type { Tab } from './Tabs.svelte';
 export { default as Tabs } from './Tabs.svelte';
+// Re-export types from types file
+export type { FilterOption, Tab } from './types.js';
